// src/assets/build/rappid.min.d.ts
        interface Options<T extends Cell> extends mvc.ViewOptions<T> {
            id?: string
        }
// ...
export namespace layout {

    export namespace DirectedGraph {

        interface Edge {
            minLen?: number;
            weight?: number;
            labelpos?: 'l' | 'c' | 'r';
            labeloffset?: number;
            width?: number;
            height?: number;
        }

        interface Node {
            width?: number;
            height?: number;
        }

        interface LayoutOptions {
            rankDir?: 'TB' | 'BT' | 'LR' | 'RL';
            ranker?: 'network-simplex' | 'tight-tree' | 'longest-path';
            nodeSep?: number;
            edgeSep?: number;
            rankSep?: number;
            marginX?: number;
            marginY?: number;
            resizeCluster?: boolean;
            clusterPadding?: dia.Padding;
            setPosition?: (element: dia.Element, position: dia.BBox) => void;
            setVertices?: boolean | ((link: dia.Link, vertices: dia.Point[]) => void);
            setLabels?: boolean | ((link: dia.Link, position: dia.Point, points: dia.Point[]) => void);
            debugTiming?: boolean;
            exportElement?: (element: dia.Element) => Node;
            exportLink?: (link: dia.Link) => Edge;
            // deprecated
            setLinkVertices?: boolean;
        }

        export function layout(graph: dia.Graph | dia.Cell[], opt?: LayoutOptions): g.Rect;
    }
}
// ...
export namespace ui {
// ...
    namespace Stencil {
// ...
        export interface Group {
            label: string;
            index: number;
            closed?: boolean;
            height?: number;
            layout?: boolean | layout.GridLayout.Options | { [key: string]: any };
        }
    }
// ...
}
// ...
export namespace layout {

    namespace ForceDirected {
        export interface Options {
            graph: dia.Graph;
            width: number;
            height: number;
            gravityCenter: dia.Point;
            charge?: number;
            linkDistance?: number;
            linkStrength?: number;

        }
    }

    class ForceDirected extends Backbone.Model {

        constructor(opt: ForceDirected.Options);

        start(): void;

        step(): void;
    }

    namespace GridLayout {

        export interface Options {
            centre?: boolean;
            dx?: number;
            dy?: number;
            resizeToFit?: boolean;
            marginX?: number;
            marginY?: number;
            columns?: number;
            columnWidth?: 'compact' | 'auto' | number | string;
            rowHeight?: 'compact' | 'auto' | number | string;
            deep?: boolean;
            parentRelative?: boolean;
        }

        export function layout(graphOrCells: dia.Graph | Array<dia.Cell>, opt?: Options): void;
    }

    namespace TreeLayout {

        interface AttributeNames {
            'siblingRank'?: string;
            'direction'?: string;
            'margin'?: string;
            'offset'?: string;
            'prevSiblingGap'?: string;
            'nextSiblingGap'?: string;
            'firstChildGap'?: string;
        }

        type Direction = 'L' | 'R' | 'T' | 'B' | 'BR' | 'BL' | 'TR' | 'TL';

        type DirectionRule = (rule: [Direction, Direction]) => (direction: Direction) => Direction;

        export interface DirectionRules {
            rotate: DirectionRule;
            flip: DirectionRule;
            straighten: DirectionRule;
        }

        export interface Options {
            graph: dia.Graph;
            gap?: number;
            parentGap?: number;
            siblingGap?: number;
            firstChildGap?: number;
            direction?: Direction;
            directionRule?: DirectionRule;
            updatePosition?: null | ((element: dia.Element, position: dia.Point, opt?: { [key: string]: any }) => void);
            updateVertices?: null | ((link: dia.Link, vertices: Array<dia.Point>, opt?: { [key: string]: any }) => void);
            updateAttributes?: null | ((layoutArea: any, root: dia.Element, rootLink: dia.Link, opt: { [key: string]: any }) => void);
            filter?: null | ((children: dia.Element[], parent: dia.Element | null, opt: { [key: string]: any }) => dia.Element[]);
            attributeNames?: AttributeNames;
        }
    }

    class TreeLayout extends Backbone.Model {

        constructor(opt: TreeLayout.Options);

        layout(opt?: { [key: string]: any }): this;

        layoutTree(root: dia.Element, opt?: { [key: string]: any }): this;

        static directionRules: TreeLayout.DirectionRules
    }
}